requirements-docker.txt
numpy~=1.20.2
paho-mqtt~=1.5.1